requirements.txt
django>=4.1 
uvicorn 
gunicorn
tzdata
azure-cosmos
aiohttp
djangorestframework

